requirements.txt
Flask[dotenv]==2.3.3
python-dotenv==1.0.0
groq==0.31.0
gunicorn==21.2.0
